backend/requirements-fastapi.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
python-multipart==0.0.12

# Async HTTP
httpx==0.27.2
aiohttp==3.10.10

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
psycopg2-binary==2.9.10

# Redis
redis[hiredis]==5.2.0
aioredis==2.0.1

# Firebase (existing)
firebase-admin==6.5.0
google-cloud-firestore==2.21.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Providers
openai==1.51.0
google-generativeai==0.8.2
deepgram-sdk==3.7.0

# Queue & Workers
celery==5.4.0
# rabbitmq==0.2.1

# Monitoring & Logging
sentry-sdk[fastapi]==2.19.0
prometheus-client==0.21.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
python-docx==1.1.2

